requirements.txt
streamlit
opencv-python-headless
deepface
tensorflow==2.12.0
numpy
pillow
